crisil/requirements.txt
Django~=5.1.7
djangorestframework~=3.16.0
psycopg2~=2.9.10